chore: encapsulate question askers and model.
